refactor: separate BoardList logic into hook, types, and queries files
