+++ 03-maincamp-triptalk/03-CNA/src/components/board-list/queries.ts
@@ -1,0 +1,19 @@
+import { gql } from "@apollo/client";
+
+
+export const FETCH_BOARDS = gql`
+  query fetchBoards($page: Int) {
+    fetchBoards(page: $page) {
+      _id
+      title
+      writer
+      createdAt
+    }
+  }
+`;
+
+export const DELETE_BOARD = gql`
+  mutation deleteBoard($boardId: ID!) {
+    deleteBoard(boardId: $boardId)
+  }
+`;
