+++ 03-maincamp-triptalk/03-CNA/src/components/board-list/types.ts
@@ -1,0 +1,6 @@
+export interface BoardListItem {
+  _id: string;
+  title: string;
+  writer: string;
+  createdAt: string;
+}
